Moved sessions tests to C++ SDK repo
