+++ ydb/public/sdk/cpp/tests/integration/ya.make
@@ -2,4 +2,6 @@ RECURSE(
     basic_example
     bulk_upsert
     server_restart
+    sessions
+    sessions_pool
 )
