+++ ydb/services/ydb/ya.make
@@ -39,8 +39,6 @@ END()
 
 RECURSE_FOR_TESTS(
     backup_ut
-    sdk_sessions_ut
-    sdk_sessions_pool_ut
     table_split_ut
     ut
 )
